Fix progress indicator.
The original DOS version used the conio.h functions to move the cursor
around on screen, and we have stub versions of the conio functions we
use on other platforms. Since we're not going to target DOS any more,
let's fix the progress indicator to use backspace characters to move
the cursor back.