Refactor how WAD files are opened.
This moves responsibility to `Compress()`, `Uncompress()` and
`ListEntries()` to open the WAD file, meaning that these functions now
both open and close the file. This makes the ownership clearer; we can
also eliminate a global variable.

A minor bug fixed here is that the old WAD files are now opened as
read-only, as none of this code actually needs write permission to
these files (we only ever generate a new file).