Fix "scrolling walls bug".
Scrolling wall linedefs work by updating the x offset associated with
the linedef while the game is running. If sidedefs are packed this means
that multiple linedefs may be updating the same sidedef (or a wall might
be scrolling that shouldn't be). We can fix this by identifying sidedefs
that belong to a scrolling linedef and never merging them with any other
sidedefs.

Fixes #2.