Rewrite WAD merging code.
We can do this much more efficiently by calculating the SHA1 hash of
each lump we write and keeping a table of the hashes of each lump we've
written so far. As part of this rewrite, we also get rid of the "suggest
mode" code, which is a relic left over from the earliest versions of the
project.