Rewrite sidedef packing.
We can reuse the sorting code to reimplement sidedef packing using the
same sorted map that we generate for merging WAD lumps.

This improves the performance of the sidedef packing algorithm. The old
approach used a naive O(n^2) algorithm. On modern machines, this was
fast enough for most levels; however, I found that for extremely large
levels, it could still be very slow. By using the sorting code the
complexity reduces to O(n log n).

The example I have been testing with is Eviternity MAP32, which
previously took O(a minute) to compress. With this change in place, it
takes < 30s to compress the entire WAD.

This fixes #6.