Optimization for sort code.
With wadptr the whole point is to build tables where the values are
repeated (reused pointers), so it's very likely that a sorted list can
include long runs of the same value. This presents a problem because it
triggers the degenerate case with quicksort where each recursion of a
long block only reduces its length by 1. So detect when we encounter
such a block and just skip the recursion step - such a block is already
sorted.