Fix fallback handling of large levels.
commit 920aa503d9153147f1 added a check for length of the SIDEDEFS lump
to make sure we never generate one that exceeds the vanilla limit.
However, we were still writing out the modified (decompressed) LINEDEFS
lump. So we need to preserve the original LINEDEFS and fall back to it
too if we can't use the new SIDEDEFS.

With this change in place, we can now add a test case to check this
continues to work as intended in the future. Here we're using Eviternity
MAP32, an enormous level that can only be stored with compressed
SIDEDEFS and BLOCKMAP (and even then, only with extended limits).