Write columns in order to optimize suffix matching.
When checking for matching columns we scan through those columns that
have already been written. This means that a suffix match will not work
if the short column is written before the long one. As a minor
optimization we can therefore write columns in order of decreasing size;
this will ensure that any opportunities for suffix matching are taken.