Change approach for sorting identical values.
The previous approach worked but we can do better by eliminating the
possibility of identical comparisons completely. The elements arrays
always consist of sets of unique values, so we can just compare them
as a fallback when the comparison function returns zero. This has the
advantage of producing a consistent ordering and I suspect (but haven't
tested) that it is probably more efficient.