Sort block sizes to optimize compression.
By building the blockmap lump in order of decreasing size, we can do
better prefix matching in some cases. The savings are only very small
with the usual block format because of the 0000 prefix to block lists,
but for levels that don't use that prefix we might be able to save a
bit more space.

We also include another minor optimization by always moving the very
longest block to the end of the lump; in the case of levels that are
just on the edge of overflowing the vanilla limits, this might just
save us.

This fixes #10.