Eliminate last uses of wadglobal.
This is relatively straightforward now; we just replace all uses with a
local variable in each function. There are a couple of minor tweaks
needed in that we need to initialize the wad_file_t struct contents to
zero and free the WAD directory at the end of processing for each file.