Don't set "x" fopen mode on Windows.
We get open failures on Windows when trying to write any file (#12), I
believe because the Win32 version of fopen() doesn't support the 'x'
exclusive mode and explicitly does not like having that character in the
mode string. While we're here, add a perror() call to dump some
additional information about the fopen error.

This fixes #12. Thanks to @Kroc for reporting.