Document why SIDEDEFS lumps sometimes grow in size.
wadptr's behavior with sidedefs on special lines is more cautious than
other tools and this is by design. However, the result is that SIDEDEFS
lumps can actually grow in size if repacking a level that was already
packed by another tool.

This is not a bug but it is counterintuitive, so add it to the manual
and detect when it happens so that we can print a message explaining
that it is deliberate behavior.

Fixes #20. Thanks @Kroc for reporting planisf2.wad that exhibits this
behavior.