More accurate detection of squashed graphics.
The column pointer can also refer to a suffix of another column rather
than being exactly equal. To detect this more efficiently, we can use
the sorting code to iterate through columns in order of offset and
identify when columns have overlapping data. This is also more
efficient than the previous O(n^2) algorithm.

With this in place, some graphics lumps are now identified as squashed
that were not previously. For example, with a compressed version of
doom2.wad, PAINM0 and SARGH1 are both identified as squashed,
presumably because the squashing algorithm was able to detect suffix
matches.