Break pthreads dependency, use `gettimeofday`
Commit 5c0f1724e039df8 added a dependency on the pthreads library (as
`clock_gettime()` is part of the pthreads API). On Windows this adds
an unwanted DLL dependency on libwinpthread (the MingW implementation of
phtreads). We can break this dependency by using `gettimeofday`; this is
a wall-time API so it should satisfy the requirement that the original
commit was aiming for.

Fixes #28. Thanks to @Anonysussy and @happygreenfairy for their bug
reports.