Add sector tag range for special effects
We are deliberately very cautious and avoid merging sidedefs attached to
special lines (for reasons already documented). However, there are some
cases where it is actually desirable to have this behavior - for
example, by making two scrolling linedefs share a sidedef so that it
scrolls faster than usual.

To support this use case, this change adds support for a magic sector
tag range (9700-9799 decimal) that wadptr now recognizes. If two
linedefs have the same sector tag and it is in this range, their
sidedefs can be merged, bypassing the usual behavior. Level authors will
therefore be able to get this behavior if they desire.

This fixes #24. Thanks to @kitchen-ace for the suggestion.